Automatically run AD tests on github
This is far from optimal, yet.

Ideally we would install the AD server on a Github Windows runner.
However this seems not to be possible without intermediate reboots which
we can't do on a Github runner.

This instead opts to use a Vagrant managed virtual machine on top of a
MacOSX runner.

The workflow makes use of the cache action which in theory would cache
the Vagrant box. Unfortunately the box is about 8GB and the cache is
limited to 5GB, so no caching happens.

If anyone has better ideas on how to run actual real tests against an
autoprovisioned AD server, please let me know.